requirements.txt
streamlit==0.84.0
numpy==1.20.1
pandas==1.0.2
opencv-contrib-python==4.5.2.54
pytorch==1.8.1
